+++ shared/src/lib.rs
@@ -4,14 +4,16 @@ use bitcoin_hashes::Hash as BitcoinHash;
 use sha2::{Digest, Sha512_256};
 
 use bitcoin::consensus::Encodable;
-use bitcoin::key::Keypair;
-use bitcoin::{BlockHash, Transaction};
+use bitcoin::key::{Keypair, UntweakedPublicKey};
+use bitcoin::script::{Builder, PushBytes};
+use bitcoin::{
+    BlockHash, ScriptBuf, TapNodeHash, TapTweakHash, Transaction, WitnessVersion, XOnlyPublicKey,
+};
+use k256::schnorr;
 use k256::PublicKey;
 
-use musig2::{
-    k256, AggNonce, FirstRound, KeyAggContext, PartialSignature, PubNonce, SecNonce,
-    SecNonceSpices, SecondRound,
-};
+use musig2::k256::elliptic_curve::sec1::ToEncodedPoint;
+use musig2::{k256, KeyAggContext};
 
 pub const UTREEXO_TAG_V1: [u8; 64] = [
     0x5b, 0x83, 0x2d, 0xb8, 0xca, 0x26, 0xc2, 0x5b, 0xe1, 0xc5, 0x42, 0xd6, 0xcc, 0xed, 0xdd, 0xa8,
@@ -73,4 +75,49 @@ pub fn sort_pubkeys(pubkeys: &mut Vec<PublicKey>) {
 
 pub fn sort_keypairs(kp: &mut Vec<Keypair>) {
     kp.sort_by(|a, b| a.public_key().serialize().cmp(&b.public_key().serialize()));
+}
+pub fn new_p2tr(internal_key: UntweakedPublicKey, merkle_root: Option<TapNodeHash>) -> ScriptBuf {
+    let output_key = tap_tweak(internal_key, merkle_root);
+    // output key is 32 bytes long, so it's safe to use `new_witness_program_unchecked` (Segwitv1)
+    new_witness_program_unchecked(WitnessVersion::V1, output_key.serialize())
+}
+
+fn new_witness_program_unchecked<T: AsRef<PushBytes>>(
+    version: WitnessVersion,
+    program: T,
+) -> ScriptBuf {
+    let program = program.as_ref();
+    debug_assert!(program.len() >= 2 && program.len() <= 40);
+    // In segwit v0, the program must be 20 or 32 bytes long.
+    debug_assert!(version != WitnessVersion::V0 || program.len() == 20 || program.len() == 32);
+    Builder::new()
+        .push_opcode(version.into())
+        .push_slice(program)
+        .into_script()
+}
+
+fn tap_tweak(internal_key: UntweakedPublicKey, merkle_root: Option<TapNodeHash>) -> XOnlyPublicKey {
+    let tweak = TapTweakHash::from_key_and_tweak(internal_key, merkle_root).to_scalar();
+
+    let pub_bytes = internal_key.serialize();
+    let pub_key: k256::PublicKey = schnorr::VerifyingKey::from_bytes(&pub_bytes)
+        .unwrap()
+        .into();
+    let pub_point = pub_key.to_projective();
+
+    let tweak_bytes = &tweak.to_be_bytes();
+    let tweak_point = k256::SecretKey::from_bytes(tweak_bytes.into())
+        .unwrap()
+        .public_key()
+        .to_projective();
+
+    let tweaked_point = pub_point + tweak_point;
+    let compressed = tweaked_point.to_encoded_point(true);
+    let x_coordinate = compressed.x().unwrap();
+
+    let ver_key = schnorr::VerifyingKey::from_bytes(&x_coordinate).unwrap();
+
+    let pubx = XOnlyPublicKey::from_slice(ver_key.to_bytes().as_slice()).unwrap();
+
+    pubx
 }
